网页上传处理逻辑

+++ src/main/java/tools/Txt.java
@@ -5,9 +5,10 @@ import java.io.*;
 public class Txt {
     public static void WriteDictionary(String text,String filePath,Boolean append,String charSet)
     {
-        FileWriter fw = null;
         try {
             File f=new File(filePath);
+            if(!f.exists()  && !f.isDirectory())
+                return ;
             FileOutputStream writerStream = new FileOutputStream(f,append);
 
             BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(writerStream, charSet));
